+++ crates/pluot_core/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod bindings;
 pub mod d3;
 pub mod params;
+pub mod render_types;
 // mod plots;
 pub mod cache;
 pub mod layer_traits;
