+++ crates/pluot_zarr/src/layers/zarr_point_layer.rs
@@ -1,6 +1,4 @@
 use std::sync::Arc;
-use pluot_core::params::PrepareResult;
-use pluot_core::params::RenderResult;
 use serde::{Deserialize, Serialize};
 use svg::node::element::Group;
 
@@ -15,6 +13,7 @@ use pluot_core::cache::{get_or_init_store, use_memo_vec_f32, use_memo_vec_i32};
 use pluot_core::two::svg::update_svg;
 use pluot_core::layer_traits::{DrawToCanvas, DrawToSvg, PreparedLayer, ViewParams, AspectRatioMode, UnitsMode, MarginParams};
 use pluot_core::layers::point_layer::{PointShapeMode, PointLayerParams, base_draw_point_layer, base_draw_point_layer_svg};
+use pluot_core::render_types::{PrepareResult, RenderResult};
 
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
